Initial commit of source files.
This program is based on a bot I wrote a few years ago to get high scores
in the single-player Flash game "Jorinapeka". I've rewritten major chunks
of it to do an alpha-beta search of the game space for the two-player
Flash game "Barca" (playbarca.com/?page_id=250).

This has not yet been tested. It will work only on a system with
BOTH of the following programs/scripts installed and in your path:

    (1) xdotool, for generating clicks programmatically

    (2) Either ImageMagick "import" (Linux) or
        Apple "screencapture" (OS X)

Also, you'll need libpng (sudo apt-get libpng-dev) in order to
compile it. Since both libpng and xdotool are hard-to-impossible
to install on OS X, I'm planning to test it on Ubuntu Linux first.