Get the bot working (and reliably beating playbarca.com!)
The bot now beats playbarca.com 100% of the time, often in
fewer than 15 moves.

The bot logs the length of each game, and whether it was a
win or a loss, to /tmp/barca.log.

When the bot crashes, the last screenshot it took will
have been saved as /tmp/jorina.png (a vestige of my old
Jorinapeka bot); you can sometimes use this to debug the
problem.