Trivial improvements.
The reason I have to keep tweaking square_weight's switch
statement is because the green circles of neighboring squares
can affect the number of light and dark pixels counted.

Also, if we get a "game over" screenshot twice in a row, that
just means that we need to wait a bit longer for the game to
reset itself. (Hopefully.)