Rename ME and YOU to WHITE and BLACK. [untested]
Should be no change in behavior. This patch just renames
ME/YOU/my_pieces/your_pieces to WHITE/BLACK/white_pieces/black_pieces.
This will make the next step a lot less confusing, when "I" (the AI)
will become able to play for either side.

Right now the MOVE_FOR() macro is defined to return true for WHITE
only. The next step is to make it dependent on a command-line option;
and then we should be done!