Allow the AI to play either side.
The "./barca" program now requires at least one of the options
"--black" and "--white", which tell it which side(s) to play.
For example, you can tell it "./barca --black" (after switching
the Flash game to human-versus-human mode), and then you should
be able to play a full game of Barca against the program,
essentially replacing the Flash game's AI with this one.

One major remaining issue is that the green arrows confuse the
input routine: if you click on an elephant or mouse along the
edge of the board, then each of that row/column's squares will
get a green arrow, changing their pixel count and making the
program think that the row isn't there anymore (causing an
assert-fail). The right thing for me to do is count green pixels
as white (or blue) for this purpose, or (better!) filter on the
square shape of the region rather than its exact pixel count.