Recognize squares by their 47x47 bounding box. [untested]
I've thrown in a minimum-population term too, just in case the main
check catches other elements such as circles and pieces.

This should prevent the assert-fail when the human player clicks
on a piece at the edge of the board and the green arrows appear
all throughout that row; those regions will now be correctly
recognized as "square" despite having arrow-shaped holes in them.