Fix a bug in the case that a scared piece is trapped. [untested]
I had assumed that in the following situation, with a black mouse at A10
surrounded by enemies,

 m.E..
 .L...
 .....
 E...

the mouse would be unable to move. It turns out that at least as far as
the Flash implementation is concerned, the mouse *is* allowed to move to
any of those three threatened squares (assuming of course that no other
scared black piece is able to move out of danger this turn). However, just
because the mouse is allowed to *stay* in danger doesn't imply that any
non-scared black piece is allowed to move *into* danger this turn.