Remove some ancient dead code from ImageFmtc.c.
I have no idea what this code was originally, back in 2004, intended to
be used for. Obviously there's more "unused" code in ImageFmtc.c, as far
as the Barca bot is concerned; but this static function was literally
unreachable, as Clang helpfully pointed out to me.

No change in behavior.